Delete files in /etc, /usr and some dotfiles
The files in /etc and /usr directories are not dotfiles, but rather
system configuration files. Because of that they should not be placed in
the dotfiles repo. I made them gists instead.

I also removed some files in the .config folder that I no longer use.
I turned the `fonts.conf` and the `sxhkdrc` files into gists as well
just in case I need them in the future.