Add .local/bin scripts and suckless config files.
The following files were previously in the `suckless` repo, but it makes
more sence to have them in the `dotfiles` repo.
- .Xresources
- .config/picom/picom.conf
- .xsession

The .xsession file is now more selective and will only run certain
programs under a DWM session.

The script `dmenu_playerctl` helps picking a player to control, and
passes playerctl commands to the chosen player. Playerctl relies on the
mpris protocol, which was added to ncmpcpp and mpv thru plugins.

I also added a couple of scripts fro exporting and importing gpg key
pairs.