Cals: Make code more modular and cache book titles
The `cals` script for dmenu is now more modular by placing chunks of
code into functions. The script now caches retrieved book titles, making
subsequent script executions faster. The script updates the cache every
time the calibre library is modified.

The script is also more robust because it looks for the "Calibre
Library" folder if its path is not specified. The search is limited to a
depth of 3 subdirectories in the home directory to maintain speed.