Add passb, a password backup script.
Passb is a script that uses `gpgtar` to encrypts `pass` passwords and `gpg`
private/public keys into a single file for easy backup.

The script can also import and decrypt the backups it has created.
Additionally, it reports new, modified, and deleted passwords. These reports
are cleared every time a backup is made. It also shows a reminder telling you
to backup the backup.

Purpose

The purpose of this script is to stop being dependent on password
managers, since they are valuable targets for hackers.

Features

Passb is a wrapper around the `pass` program, so it can do everything that
`pass` does in addition to the following:

* List passwords that are new, modified, or deleted.
* Create encrypted backups of all passwords and private/public keys.
* Export the backups into a given directory.
* Decrypt the backups into a given directory.
* Easily import backups or GPG encrypted files.
* Easily delete backups or GPG encrypted files.

Possible improvements

* Use `shred` instead of `rm` to delete files entirely.