Add scripts for switching audio and word lookup.
The `s` script stands for StarDict. It searches a word and opens its
definition in a terminal.

The `switch_audio_source` script switches between the audio coming from
the headphones and the speakers. Using this command is more convenient
than physically plugging and unplugging the headphones or opening some
other program to switch the audio source.