Add nvim aliases, plugins, and functions.
This commit adds functions for toggling automatic program compilation,
debug builds, and code formatting.

The code for compiling and running code was refactored to make it more
maintainable.

A new function for loading templates was also added.

An indentation plugin was added to make heavily-indented code easier to
read, while a tagbar plugin was added to facilitate moving across the
codebase.